pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lab-memory"
version = "0.1.0"
description = "Personal research knowledge retrieval system for lab reports"
requires-python = ">=3.11"
dependencies = [
    "python-pptx==1.0.2",
    "PyMuPDF==1.25.3",
    "sentence-transformers==3.4.1",
    "chromadb==0.6.3",
    "mcp[cli]==1.8.0",
    "anthropic==0.49.0",
    "click==8.1.8",
    "pyyaml==6.0.2",
    "tenacity==9.0.0",
]

[project.scripts]
lab-memory = "lab_memory.cli:cli"

[tool.setuptools.packages.find]
include = ["lab_memory*"]
